+++ app/ent/profile/where.go
@@ -10,47 +10,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.Profile {
+func ID(id int8) predicate.Profile {
 	return predicate.Profile(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.Profile {
+func IDEQ(id int8) predicate.Profile {
 	return predicate.Profile(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.Profile {
+func IDNEQ(id int8) predicate.Profile {
 	return predicate.Profile(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.Profile {
+func IDIn(ids ...int8) predicate.Profile {
 	return predicate.Profile(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.Profile {
+func IDNotIn(ids ...int8) predicate.Profile {
 	return predicate.Profile(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.Profile {
+func IDGT(id int8) predicate.Profile {
 	return predicate.Profile(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.Profile {
+func IDGTE(id int8) predicate.Profile {
 	return predicate.Profile(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.Profile {
+func IDLT(id int8) predicate.Profile {
 	return predicate.Profile(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.Profile {
+func IDLTE(id int8) predicate.Profile {
 	return predicate.Profile(sql.FieldLTE(FieldID, id))
 }
